TEST: Add test_main
