+++ test/BarAndHinge_test.cpp
@@ -1,8 +1,5 @@
+#include "test_main.cpp"
 #include "BarAndHinge.h"
-
-#include "gtest/gtest.h"
-#include "gtest-mpi-listener.hpp"
-#include "ArraysMatch.h"
 
 #include <math.h>
 #include "State.h"
@@ -56,19 +53,4 @@ TEST(BarAndHingeTest, TestMin) {
   Vector x5 = min.minimise(s5);
   Vector x5_expected = {0,0,0.5-dz, 0,-sqrt(2),0.5+dz, 0,sqrt(2),0.5+dz, 0,0,0.5-dz};
   EXPECT_TRUE(ArraysNear(x5, x5_expected, 1e-4));
-}
-
-
-int main(int argc, char** argv) {
-  ::testing::InitGoogleTest(&argc, argv);
-  MPI_Init(&argc, &argv);
-  mpi.getSizeRank(MPI_COMM_WORLD);
-
-  // Add an MPI listener (https://github.com/LLNL/gtest-mpi-listener)
-  ::testing::AddGlobalTestEnvironment(new GTestMPIListener::MPIEnvironment);
-  ::testing::TestEventListeners& listeners = ::testing::UnitTest::GetInstance()->listeners();
-  ::testing::TestEventListener *l = listeners.Release(listeners.default_result_printer());
-  listeners.Append(new GTestMPIListener::MPIWrapperPrinter(l, MPI_COMM_WORLD));
-
-  return RUN_ALL_TESTS();
 }
