+++ test/Fire_test.cpp
@@ -1,8 +1,5 @@
+#include "test_main.cpp"
 #include "Fire.h"
-
-#include "gtest/gtest.h"
-#include "gtest-mpi-listener.hpp"
-#include "ArraysMatch.h"
 
 #include <math.h>
 #include "State.h"
@@ -69,19 +66,4 @@ TEST(FireTest, TestSingleProc) {
   } else {
     EXPECT_TRUE(result.empty());
   }
-}
-
-
-int main(int argc, char** argv) {
-  ::testing::InitGoogleTest(&argc, argv);
-  MPI_Init(&argc, &argv);
-  mpi.getSizeRank(MPI_COMM_WORLD);
-
-  // Add an MPI listener (https://github.com/LLNL/gtest-mpi-listener)
-  ::testing::AddGlobalTestEnvironment(new GTestMPIListener::MPIEnvironment);
-  ::testing::TestEventListeners& listeners = ::testing::UnitTest::GetInstance()->listeners();
-  ::testing::TestEventListener *l = listeners.Release(listeners.default_result_printer());
-  listeners.Append(new GTestMPIListener::MPIWrapperPrinter(l, MPI_COMM_WORLD));
-
-  return RUN_ALL_TESTS();
 }
