+++ test/PFWetting_test.cpp
@@ -1,12 +1,8 @@
+#include "test_main.cpp"
 #include "PFWetting.h"
-
-#include "gtest/gtest.h"
-#include "gtest-mpi-listener.hpp"
-#include "ArraysMatch.h"
 
 #include "State.h"
 #include "utils/vec.h"
-#include "utils/mpi.h"
 
 using namespace minim;
 
@@ -162,19 +158,4 @@ TEST(PFWettingTest, TestNFluid) {
   EXPECT_FLOAT_EQ(pot.nFluid, 1);
   pot.setNFluid(3);
   EXPECT_FLOAT_EQ(pot.nFluid, 3);
-}
-
-
-int main(int argc, char** argv) {
-  ::testing::InitGoogleTest(&argc, argv);
-  MPI_Init(&argc, &argv);
-  mpi.getSizeRank(MPI_COMM_WORLD);
-
-  // Add an MPI listener (https://github.com/LLNL/gtest-mpi-listener)
-  ::testing::AddGlobalTestEnvironment(new GTestMPIListener::MPIEnvironment);
-  ::testing::TestEventListeners& listeners = ::testing::UnitTest::GetInstance()->listeners();
-  ::testing::TestEventListener *l = listeners.Release(listeners.default_result_printer());
-  listeners.Append(new GTestMPIListener::MPIWrapperPrinter(l, MPI_COMM_WORLD));
-
-  return RUN_ALL_TESTS();
 }
