+++ test/test_main.cpp
@@ -1,0 +1,19 @@
+#include "gtest/gtest.h"
+#include "gtest-mpi-listener.hpp"
+#include "ArraysMatch.h"
+#include "utils/mpi.h"
+
+
+int main(int argc, char** argv) {
+  ::testing::InitGoogleTest(&argc, argv);
+  MPI_Init(&argc, &argv);
+  minim::mpi.getSizeRank(MPI_COMM_WORLD);
+
+  // Add an MPI listener (https://github.com/LLNL/gtest-mpi-listener)
+  ::testing::AddGlobalTestEnvironment(new GTestMPIListener::MPIEnvironment);
+  ::testing::TestEventListeners& listeners = ::testing::UnitTest::GetInstance()->listeners();
+  ::testing::TestEventListener *l = listeners.Release(listeners.default_result_printer());
+  listeners.Append(new GTestMPIListener::MPIWrapperPrinter(l, MPI_COMM_WORLD));
+
+  return RUN_ALL_TESTS();
+}
